% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonostatics-package.R
\docType{package}
\name{harmonostatics-package}
\alias{harmonostatics}
\alias{harmonostatics-package}
\title{harmonostatics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/homeymusic/a_field_theory_of_musical_harmony}
  \item Report bugs at \url{https://github.com/homeymusic/a_field_theory_of_musical_harmony/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
