% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\title{Harmony}
\usage{
harmony(x, home, name = NULL)
}
\arguments{
\item{x}{A note or chord expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{name=NULL}{An optional custom name for the note or chord}
}
\value{
A tibble with semitone, intervallic_name, name, affinity, brightness and potential energy
}
\description{
Provides the musical harmony metrics of a note or chord.
}
\examples{
harmony(4,0) # provides the harmony metrics of the Major 3rd
harmony(8,12) # provides the harmony metrics of the inverted Major 3rd (Minor 6th with Octave)

harmony(x=c(0,4,7),home=0, name="C Major Triad") # provides the harmony metrics of a C Major Triad
harmony(x=c(12,8,5),home=12, name="Inverted C Major Triad") # provides the harmony metrics of the inverted C Major Triad

}
