% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{homey_plot_progression}
\alias{homey_plot_progression}
\title{Homey Plot Progression}
\usage{
homey_plot_progression(
  x,
  y,
  home,
  columns,
  unlist = FALSE,
  include_names = TRUE,
  symmetrical = TRUE,
  x_expansion_mult = 0.6,
  title = NULL,
  include_path = FALSE,
  y_lim_max = NA_real_,
  y_lim_min = NA_real_
)
}
\arguments{
\item{x}{A list of notes or chords expressed as an interval integers or vector of interval integers}

\item{y}{A list of notes or chords expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{columns}{A vector of 4 named columns to plot: x, y, size and color}

\item{unlist}{A logical TRUE or FALSE to plot points individually or as one point}

\item{include_names}{Include the names from the list x on the plot}

\item{symmetrical}{Center the plot horizontally}

\item{x_expansion_mult}{Leave space for the labels within the chart}

\item{title}{An optional title for the plot}

\item{include_path}{include line representing ordered path among chords or notes}

\item{y_lim_max}{Set the max vertical value for the plot, default is NA_real_}

\item{y_lim_min}{Set the min vertical value for the plot, default is NA_real_}
}
\value{
Generates the requested scatter plot and returns TRUE
}
\description{
Provides scatter plots of musical potential energy of a note or chord with the homey theme.
}
