% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_energy.R
\name{potential_energy}
\alias{potential_energy}
\title{Potential Energy}
\usage{
potential_energy(from, to, home)
}
\arguments{
\item{from}{A starting note or chord expressed as a harmony tibble}

\item{to}{An ending note or chord expressed as a harmony tibble}

\item{home}{The home note of the musical context}
}
\value{
The potential energy difference between the two chords or notes
}
\description{
Provides the potential energy between two chords or notes
}
