% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression.R
\name{progression}
\alias{progression}
\title{Progression}
\usage{
progression(from, to, home = NULL, name = NULL)
}
\arguments{
\item{from}{A starting note or chord expressed as a harmony tibble}

\item{to}{An ending note or chord expressed as a harmony tibble}

\item{home}{The home note of the current musical context}

\item{name}{A name to describe the progression}
}
\value{
A tibble with information about the change in harmony from to
}
\description{
Provides information about a progression between two chords or notes
}
