% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonostatics-package.R
\docType{package}
\name{harmonostatics-package}
\alias{harmonostatics}
\alias{harmonostatics-package}
\title{harmonostatics: Harmonostatics}
\description{
Harmonostatics allows researchers and music makers to explore a static field theory of musical harmony.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/homeymusic/a_field_theory_of_musical_harmony}
  \item Report bugs at \url{https://github.com/homeymusic/a_field_theory_of_musical_harmony/issues}
}

}
\author{
\strong{Maintainer}: Brian McAuliff Mulloy \email{brian@homeymusic.com} (\href{https://orcid.org/0000-0002-1803-3172}{ORCID})

}
\keyword{internal}
