% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\title{Harmony}
\usage{
harmony(x, home = NULL, name = NULL)
}
\arguments{
\item{x}{A note or chord expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{name}{A custom name for the note or chord}
}
\value{
A tibble with semitone, intervallic_name, name, affinity, brightness and harmony.magnitude
}
\description{
Provides the musical harmony metrics of a note or chord.
}
