% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{homey_plot_harmony}
\alias{homey_plot_harmony}
\title{Homey Plot Harmony}
\usage{
homey_plot_harmony(
  x,
  home = NULL,
  columns,
  unlist = FALSE,
  include_names = TRUE,
  title = NULL,
  pascal_triangle = FALSE,
  repel_labels = FALSE,
  max_overlaps = Inf,
  x_expansion_mult = 0.6,
  include_path = FALSE
)
}
\arguments{
\item{x}{A list of notes or chords expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{columns}{A vector of 2 column names to plot: semitone, affinity, brightness and magnitude}

\item{unlist}{A logical TRUE or FALSE to plot points individually or as one point}

\item{include_names}{Include the names from the list x on the plot}

\item{title}{An optional title for the plot}

\item{pascal_triangle}{Plot with triangular numbers as vertical gridlines}

\item{repel_labels}{Space labels apart}

\item{max_overlaps}{for repel labels}

\item{x_expansion_mult}{add add padding horizontally, 0.6 is default}

\item{include_path}{include line representing ordered path among chords or notes}
}
\value{
Generates the requested scatter plot and returns TRUE
}
\description{
Provides scatter plots of musical harmony metrics of a note or chord with the homey theme.
}
