% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_energy_plots.R
\name{homey_plot_potential_energy}
\alias{homey_plot_potential_energy}
\title{Homey Plot Potential Energy}
\usage{
homey_plot_potential_energy(
  x,
  y,
  home,
  columns,
  unlist = FALSE,
  include_names = TRUE,
  symmetrical = TRUE,
  expansion_mult = 0.6,
  title = NULL
)
}
\arguments{
\item{x}{A list of notes or chords expressed as an interval integers or vector of interval integers}

\item{y}{A list of notes or chords expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{columns}{A vector of 2 column names to plot: semitone, affinity, brightness and magnitude}

\item{unlist=FALSE}{A logical TRUE or FALSE to plot points individually or as one point}

\item{include_names=TRUE}{Include the names from the list x on the plot}

\item{title=NULL}{An optional title for the plot}
}
\value{
Generates the requested scatter plot and returns TRUE
}
\description{
Provides scatter plots of musical potential energy of a note or chord with the homey theme.
}
