% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_energy.R
\name{potential_energy}
\alias{potential_energy}
\title{Potential Energy}
\usage{
potential_energy(x, y, home, name = NULL)
}
\arguments{
\item{x}{A note or chord expressed as an interval integers or vector of interval integers}

\item{y}{A note or chord expressed as an interval integers or vector of interval integers}

\item{home}{The home pitch expressed an as interval integer}

\item{name=NULL}{An optional custom name for the note or chord}
}
\value{
A tibble with semitone, intervallic_name, name, affinity, brightness, magnitude and potential energy
}
\description{
Provides the potential energy between two chords or notes
}
